Average of bearings

Average of bearings

average_bearing(theta1, theta2, average_of_opposite = NULL)

average_bearing_n(thetas)

Arguments

theta1, theta2
Bearings, expressed in degrees.

average_of_opposite
The average of opposing bearings (e.g. average of north and south) is not well-defined. If NULL, the result for opposing vectors is undefined; if "right", returns theta1 + 90; if "left" then theta2 + 90. Can also be a single numeric to provide a specific value when the vectors point in opposite directions.

thetas
A vector of bearings.

Value

For `average_bearing`, the bearing bisecting the two bearings.

For `average_bearing_n`, the average bearing of the bearing.

Examples

average_bearing(0, 90)

#> [1] 45

average_bearing(0, 270)

#> [1] 315

average_bearing(90, 180)

#> [1] 135

average_bearing(0, 180)

#> [1] 0

average_bearing(0, 180, average_of_opposite = 3)

#> [1] 3

average_bearing(0, 180, average_of_opposite = "left")

#> [1] 270

average_bearing_n(1:179)

#> [1] 90